requirements.txt
asgiref==2.3.2
async-timeout==3.0.1
attrs==18.2.0
autobahn==18.12.1
Automat==0.7.0
beautifulsoup4==4.6.3
certifi==2018.11.29
chardet==3.0.4
constantly==15.1.0
Django==2.1.3
django-autocomplete-light==3.3.2
django-bootstrap-form==3.4
django-channels==0.7.0
django-chat==0.1.0
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-emojione==0.1.7.4
django-filter==2.0.0
django-helpdesk==0.2.10
django-js-asset==1.1.0
django-markdown-deux==1.0.5
django-precise-bbcode==1.2.10
django-vote==2.1.7
email-reply-parser==0.5.9
emojipy==3.0.5
hyperlink==18.0.0
idna==2.8
incremental==17.5.0
lxml==4.2.5
markdown2==2.3.7
oauthlib==2.1.0
Pillow==5.3.0
PyHamcrest==1.9.0
pytz==2018.7
requests==2.21.0
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.12.0
txaio==18.8.1
urllib3==1.24.1
zope.interface==4.6.0
